pyproject.toml
[tool.poetry]
name = "gpt-diffy"
version = "0.1.0"
description = "A Python CLI tool to generate Git commit messages using OpenAI's GPT technology."
authors = ["Matthew Spah <spahmatthew@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/zero-sum-seattle/gpt-diffy"
repository = "https://github.com/zero-sum-seattle/gpt-diffy"

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.9.0"
openai = "^0.27.8"
gitpython = "^3.1.35"
pyyaml = "^6.0"
[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"

